add additional runtime checks and gfx1201 fix
This commit contains three fixes:
 - increase the max. number of files at the beginning of the run to the
   max. allowed by the system
 - check for large BAR support. WE don not abort if its not available,
   but print a warning.
 - for gfx1201, do not use uncached memory at the moment.
